# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Whisper Web API
  description: |
    # Kiswahili Health AI Backend API
    
    This API provides speech-to-text transcription, conversation management, and health survey functionality for the Whisper Web application.
    
    ## Features
    - 🎤 **Speech Transcription**: Convert audio to text using OpenAI Whisper
    - 💬 **Conversation Management**: Handle multi-turn health conversations
    - 📊 **Metrics Tracking**: Real-time quality and performance metrics
    - 🌍 **Kiswahili Support**: Optimized for African language processing
    - 🔄 **WebSocket Support**: Real-time bidirectional communication
    
    ## Authentication
    Currently, this API doesn't require authentication for development purposes.
    
    ## Rate Limiting
    - 100 requests per minute per IP address
    - File uploads limited to 10MB
    
    ## WebSocket Events
    - `join-session`: Join a conversation session
    - `send-message`: Send a message to the AI
    - `message-response`: Receive AI response
    - `metrics-update`: Real-time metrics updates
    
  version: 1.0.0
  contact:
    name: Whisper Web Team
    email: support@whisper-web.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.whisper-web.com
    description: Production server

tags:
  - name: Health
    description: Health survey and conversation endpoints
  - name: Speech
    description: Speech-to-text transcription endpoints
  - name: Metrics
    description: Performance and quality metrics
  - name: WebSocket
    description: Real-time communication endpoints

paths:
  /api/health:
    get:
      tags:
        - Health
      summary: Health check endpoint
      description: Check if the API is running and healthy
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2024-01-15T10:30:00Z"
                  uptime:
                    type: number
                    description: Server uptime in seconds
                    example: 3600
                  version:
                    type: string
                    example: "1.0.0"
                  services:
                    type: object
                    properties:
                      speech:
                        type: string
                        example: "operational"
                      conversation:
                        type: string
                        example: "operational"
                      metrics:
                        type: string
                        example: "operational"

  /api/conversation/start:
    post:
      tags:
        - Health
      summary: Start a new conversation session
      description: Initialize a new health conversation session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                language:
                  type: string
                  default: "sw-KE"
                  description: Language code for the conversation
                  example: "sw-KE"
                useCase:
                  type: string
                  enum: [health, general]
                  default: "health"
                  description: Type of conversation
                  example: "health"
      responses:
        '201':
          description: Conversation session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    description: Unique session identifier
                    example: "session_123456789"
                  message:
                    type: string
                    description: Welcome message
                    example: "Karibu! Mimi ni daktari wako wa AI. Unaweza kuniambia nini kinaumiza?"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2024-01-15T10:30:00Z"

  /api/conversation/{sessionId}/message:
    post:
      tags:
        - Health
      summary: Send a message in a conversation
      description: Send a message and receive AI response
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          description: Conversation session ID
          example: "session_123456789"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: User's message
                  example: "Nina maumivu ya tumbo"
                language:
                  type: string
                  default: "sw-KE"
                  description: Language code
                  example: "sw-KE"
      responses:
        '200':
          description: AI response generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Message ID
                    example: "msg_987654321"
                  role:
                    type: string
                    enum: [user, daktari]
                    description: Message sender role
                    example: "daktari"
                  content:
                    type: string
                    description: Response content
                    example: "Pole sana kwa maumivu hayo. Je, ni mahali gani haswa mwilini unaposikia uchungu?"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2024-01-15T10:30:00Z"
                  confidence:
                    type: number
                    minimum: 0
                    maximum: 100
                    description: AI confidence score
                    example: 85.5

  /api/speech/transcribe:
    post:
      tags:
        - Speech
      summary: Transcribe audio to text
      description: Convert audio file to text using OpenAI Whisper
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audio
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file to transcribe
                language:
                  type: string
                  default: "sw"
                  description: Language code for transcription
                  example: "sw"
                model:
                  type: string
                  default: "whisper-1"
                  description: Whisper model to use
                  example: "whisper-1"
      responses:
        '200':
          description: Audio transcribed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                    description: Transcribed text
                    example: "Nina maumivu ya tumbo na homa"
                  language:
                    type: string
                    description: Detected language
                    example: "sw"
                  confidence:
                    type: number
                    minimum: 0
                    maximum: 1
                    description: Transcription confidence
                    example: 0.92
                  duration:
                    type: number
                    description: Audio duration in seconds
                    example: 3.5
                  timestamp:
                    type: string
                    format: date-time
                    example: "2024-01-15T10:30:00Z"

  /api/metrics/session/{sessionId}:
    get:
      tags:
        - Metrics
      summary: Get session metrics
      description: Retrieve performance metrics for a conversation session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          description: Session ID
          example: "session_123456789"
      responses:
        '200':
          description: Session metrics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    example: "session_123456789"
                  totalMessages:
                    type: integer
                    description: Total messages in session
                    example: 10
                  duration:
                    type: number
                    description: Session duration in milliseconds
                    example: 300000
                  averages:
                    type: object
                    properties:
                      kiswahiliAccuracy:
                        type: number
                        minimum: 0
                        maximum: 100
                        description: Average Kiswahili accuracy
                        example: 85.2
                      medicalTerminology:
                        type: number
                        minimum: 0
                        maximum: 100
                        description: Medical terminology recognition
                        example: 78.5
                      culturalSensitivity:
                        type: number
                        minimum: 0
                        maximum: 100
                        description: Cultural sensitivity score
                        example: 92.1
                      responseLatency:
                        type: number
                        description: Average response time in milliseconds
                        example: 800
                      wer:
                        type: number
                        minimum: 0
                        maximum: 1
                        description: Word Error Rate
                        example: 0.152
                  latest:
                    type: object
                    properties:
                      timestamp:
                        type: string
                        format: date-time
                        example: "2024-01-15T10:30:00Z"
                      kiswahiliAccuracy:
                        type: number
                        example: 87.0
                      medicalTerminology:
                        type: number
                        example: 80.0
                      culturalSensitivity:
                        type: number
                        example: 95.0
                      responseLatency:
                        type: number
                        example: 750
                      wer:
                        type: number
                        example: 0.148

  /api/metrics/global:
    get:
      tags:
        - Metrics
      summary: Get global metrics
      description: Retrieve global performance statistics across all sessions
      responses:
        '200':
          description: Global metrics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalSessions:
                    type: integer
                    description: Total number of sessions
                    example: 150
                  totalMessages:
                    type: integer
                    description: Total messages across all sessions
                    example: 2500
                  averageMetrics:
                    type: object
                    properties:
                      kiswahiliAccuracy:
                        type: number
                        minimum: 0
                        maximum: 100
                        description: Average Kiswahili accuracy
                        example: 83.7
                      medicalTerminology:
                        type: number
                        minimum: 0
                        maximum: 100
                        description: Medical terminology recognition
                        example: 76.2
                      culturalSensitivity:
                        type: number
                        minimum: 0
                        maximum: 100
                        description: Cultural sensitivity score
                        example: 89.5
                      responseLatency:
                        type: number
                        description: Average response time in milliseconds
                        example: 850
                      wer:
                        type: number
                        minimum: 0
                        maximum: 1
                        description: Word Error Rate
                        example: 0.163

components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid session ID"
        code:
          type: string
          description: Error code
          example: "INVALID_SESSION"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future use)

responses:
  BadRequest:
    description: Bad request
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/Error'
  
  Unauthorized:
    description: Unauthorized
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/Error'
  
  NotFound:
    description: Resource not found
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/Error'
  
  InternalServerError:
    description: Internal server error
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/Error' 